+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+*
+
+### Changed
+
+* Adapt select left padding to be consistent among other components
+
+### Fixed
+
+*
+
+## [0.5.0] - 2021-07-20
+
+### Added
+
 * Fallback image on loading error in avatar component
 * Checkbox component
 * Input component
@@ -20,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Multiple files selection with image picker
 * Date input component
 * Switcher component
+* Image carrousel component
 * Input animated component
 * Added `animateScrollBottom` prop to chat
 * `onScroll` and `onScrollBottom` props in chat component